if idri

+++ controller.py
@@ -1,5 +1,3 @@
-from __future__ import print_function
-
 import json
 import os
 import shutil
@@ -121,7 +119,8 @@ def file(protein, pdbqt):
 
 # Main method
 if __name__ == "__main__":
-    os.rmdir(os.path.join(app.instance_path, 'ligand'))
+    if os.path.isdir(os.path.join(app.instance_path, 'ligand')):
+        os.rmdir(os.path.join(app.instance_path, 'ligand'))
     os.makedirs(os.path.join(app.instance_path, 'ligand'))
     app.run()
